Added partition_id to split metadata data.
The `partition_id` is required for the merge planner to avoid
merging  splits with different `partition_id`.

Related to #1796 and #1795

This PR also changes the hasher for the SipHasher, since nothing
ensures that the implementation of the DefaultHasher won't change
in the future.

Finally, it seeds the hash using the expression itself, to avoid
a class of collisions that could happpen with a change in the partition
logic.